+++ tests/500_cli/530_scrap.py
@@ -58,6 +58,9 @@ def test_scrap_basic_success(caplog, tmp_path, settings):
 
     NOTE: This involves 2 requests to API
     """
+    manifest_path = tmp_path / "manifest.yaml"
+    poster_path = tmp_path / "cover.jpg"
+
     runner = CliRunner()
 
     result = runner.invoke(cli_frontend, [
@@ -68,9 +71,6 @@ def test_scrap_basic_success(caplog, tmp_path, settings):
         "--language", "en",
         "--write-diff",
     ])
-
-    manifest_path = tmp_path / "manifest.yaml"
-    poster_path = tmp_path / "cover.jpg"
 
     assert manifest_path.exists() is True
     assert poster_path.exists() is True
@@ -123,3 +123,46 @@ def test_scrap_diff_success(caplog, tmp_path, settings):
 
     assert result.exit_code == 0
     assert diff_path.exists() is True
+
+
+@api_allowed
+def test_scrap_basic_dry(caplog, tmp_path, settings):
+    """
+    With dry mode enabled everything is runned but nothing should be written.
+
+    NOTE: This involves 2 requests to API
+    """
+    runner = CliRunner()
+
+    poster_path = tmp_path / "cover.jpg"
+
+    result = runner.invoke(cli_frontend, [
+        "scrap",
+        SAMPLE_TV_ID,
+        str(tmp_path),
+        "--key", settings.tmdbapi_key(),
+        "--language", "en",
+        "--write-diff",
+        "--dry",
+    ])
+    assert result.exit_code == 0
+
+    assert list(tmp_path.iterdir()) == []
+
+    assert caplog.record_tuples == [
+        (
+            APPLABEL,
+            logging.INFO,
+            "TV ID: {}".format(SAMPLE_TV_ID)
+        ),
+        (
+            APPLABEL,
+            logging.INFO,
+            "Title: {}".format(SAMPLE_TV_PAYLOAD["title"])
+        ),
+        (
+            APPLABEL,
+            logging.INFO,
+            "Poster: {}".format(str(poster_path))
+        )
+    ]