+++ docs/history.rst
@@ -9,6 +9,8 @@ Version 0.6.1 - Unreleased
 
 * [collect] ``AssetStorage.store_assets()`` now returns also the asset storage
   directory along stored asset filepath list in a tuple;
+* [scrap] Added dry mode so everything is running but nothing is written or removed
+  from filesystem;
 
 
 Version 0.6.0 - 2023/07/12